// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const RoleName = {
  ADMIN: 'ADMIN',
  USER: 'USER'
} as const

export type RoleName = (typeof RoleName)[keyof typeof RoleName]


export const ReservationStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  CANCELLED: 'CANCELLED',
  EXPIRED: 'EXPIRED'
} as const

export type ReservationStatus = (typeof ReservationStatus)[keyof typeof ReservationStatus]


export const ShowtimeStatus = {
  OPEN: 'OPEN',
  CLOSED: 'CLOSED',
  CANCELLED: 'CANCELLED'
} as const

export type ShowtimeStatus = (typeof ShowtimeStatus)[keyof typeof ShowtimeStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]